For create, use the very basename of the first arg as the project name.
Essentially, we use the first argument as passed to 'create' as the
project name, but we want to handle the case where we get a path -
in that case we strip the path portion of it.

The idea is to have something similar in spirit what the 'create'
command of other tools support.